klue_re/requirements.txt
google-genai>=0.3.0
datasets>=2.14.0
pandas>=1.5.0
tqdm>=4.64.0
google-cloud-aiplatform>=1.35.0 